backend/requirements.txt
livekit-agents[deepgram,groq,silero,turn-detector]~=1.0
livekit-plugins-bey~=1.0
livekit-plugins-noise-cancellation~=0.2
livekit-api~=1.0

# HTTP API server (token + summary endpoints for the React frontend)
fastapi>=0.110
uvicorn[standard]>=0.29
groq>=0.9.0

python-dotenv>=1.0.0
aiosqlite>=0.19.0
